refactor: Move <NotificationCount/> to @chainlit/components
