+++ libs/components/src/index.ts
@@ -1,3 +1,5 @@
 export * from './buttons';
 export * from './elements';
 export * from './types';
+
+export { NotificationCount } from './NotificationCount';
