+++ libs/components/src/types/NotificationCount.tsx
@@ -1,0 +1,10 @@
+export type NotificationCountProps = {
+  count?: number | string;
+  inputProps?: {
+    id: string;
+    max?: number;
+    min?: number;
+    onChange: (event: any) => void;
+    step?: number;
+  };
+};
