+++ libs/components/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './element';
+export * from './NotificationCount';
